backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
python-dotenv
requests
passlib[bcrypt]
python-jose
pytest
httpx
python-jose[cryptography]
pygments
openai
pylint
flake8
bandit
